Added float to header

+++ include/visualise_data.hpp
@@ -4,6 +4,6 @@
 #include <vector>
 
 void printBars(const std::vector<double>& magnitudes);
-void waveformVisualiser(const std::vector<double>& magnitudes, int windowWidth,
+void waveformVisualiser(const std::vector<float>& magnitudes, int windowWidth,
                         int windowHeight);
 #endif
